chore: Move sample SOAP classes
